Separate icon creation from banner creation, since the former is obligatory

+++ make.bat
@@ -1,4 +1,5 @@
 @echo off
+:begin
 echo blargSNES -- Virtual Console Edition
 set /p title=Game Title: 
 if not exist "input\%title%\*.smc" if not exist "input\%title%\*.sfc" (
@@ -18,6 +19,14 @@ if exist "input\%title%\*.smc" (
 )
 if exist "input\%title%\*.bmp" copy /b "input\%title%\*.bmp" romfs\blargSnesBorder.bmp
 if exist "input\%title%\*.ini" copy /b "input\%title%\*.ini" romfs\blargSnes.ini
+if not exist "input\%title%\icon.png" (
+    echo ERROR: Icon image not found.
+    pause
+    call :exit
+)
+tools\convert "input\%title%\icon.png" -resize 42x42! output\tempicon.png
+tools\convert tools\icon.png output\tempicon.png -gravity center -composite "output\%title%\icon.png"
+del output\tempicon.png
 tools\bannertool makesmdh -s "%title%" -l "%title%" -p "%author%" -i "output\%title%\icon.png" -o "icon.bin"
 tools\3dstool -c -f banner.bin -t banner --banner-dir banner
 echo %title%> romfs\rom.txt
@@ -28,3 +37,11 @@ del /f /q romfs
 copy /b banner\backup banner
 rmdir /s /q banner\backup
 echo Done
+set /p choice3="Do you want to build another CIA? (Y/N): "  %=%
+if [%choice3%]==[Y] goto begin
+if [%choice3%]==[y] goto begin
+if [%choice3%]==[N] goto exit
+if [%choice3%]==[n] goto exit
+
+:exit
+exit
